fix PR changes  - remove tailwind config file  - changes indent for devcontainer.json  - use leptos_meta for title  - other misc changes
