+++ .cargo/config.toml
@@ -1,9 +1,6 @@
 [build]
 rustflags = ["--cfg=web_sys_unstable_apis"]
 
-[target.'cfg(target_arch = "wasm32")']
-rustflags = ["--cfg", "getrandom_backend=\"wasm_js\""]
-
 [target.wasm32-unknown-unknown]
 rustflags = ['--cfg', 'getrandom_backend="wasm_js"']
 
@@ -11,6 +8,7 @@ rustflags = ['--cfg', 'getrandom_backend="wasm_js"']
 [alias]
 api = "leptos serve -p dashboard"
 docker = "leptos serve -p docker"
-watch-api = "leptos watch -p dashboard"
 patr = "run -p cli --"
 prepare = "sqlx prepare --workspace"
+watch-api = "leptos watch -p dashboard"
+watch-docker = "leptos watch -p docker"
