+++ .devcontainer/devcontainer.json
@@ -1,46 +1,46 @@
 // For format details, see https://aka.ms/devcontainer.json. For config options, see the
 // README at: https://github.com/devcontainers/templates/tree/main/src/rust-postgres
 {
-  "name": "Patr API",
-  "dockerComposeFile": "docker-compose.yml",
-  "updateRemoteUserUID": true,
-  "service": "app",
-  "workspaceFolder": "/workspace",
-  "customizations": {
-    "vscode": {
-      "settings": {
-        // runs rust analyzer in different folder so that cargo build won't wait to start compiling
-        "rust-analyzer.cargo.targetDir": "./target/rust-analyzer",
-        "evenBetterToml.formatter.reorderKeys": true,
-        "files.watcherExclude": {
-          "**/target/**": true,
-          "./devcontainer/volume/**": true
-        },
-        "tailwindCSS.includeLanguages": {
-          "rust": "html"
-        }
-      },
-      "extensions": [
-        "vadimcn.vscode-lldb",
-        "rust-lang.rust-analyzer",
-        "tamasfe.even-better-toml",
-        "usernamehw.errorlens",
-        "citreae535.sparse-crates",
-        "GitHub.vscode-pull-request-github",
-        "naumovs.color-highlight",
-        "streetsidesoftware.code-spell-checker",
-        "patr-cloud.patr-theme",
-        "bradlc.vscode-tailwindcss"
-      ]
-    }
-  },
-  "forwardPorts": [
-    3000,
-    3001,
-    4000,
-    "logs:3100",
-    "tracing:16686",
-    "grafana:3000"
-  ],
-  "postAttachCommand": "sudo chown :docker /var/run/docker.sock || exit 0"
+	"name": "Patr API",
+	"dockerComposeFile": "docker-compose.yml",
+	"updateRemoteUserUID": true,
+	"service": "app",
+	"workspaceFolder": "/workspace",
+	"customizations": {
+		"vscode": {
+			"settings": {
+				// runs rust analyzer in different folder so that cargo build won't wait to start compiling
+				"rust-analyzer.cargo.targetDir": "./target/rust-analyzer",
+				"evenBetterToml.formatter.reorderKeys": true,
+				"files.watcherExclude": {
+					"**/target/**": true,
+					"./devcontainer/volume/**": true
+				},
+				"tailwindCSS.includeLanguages": {
+					"rust": "html"
+				}
+			},
+			"extensions": [
+				"vadimcn.vscode-lldb",
+				"rust-lang.rust-analyzer",
+				"tamasfe.even-better-toml",
+				"usernamehw.errorlens",
+				"citreae535.sparse-crates",
+				"GitHub.vscode-pull-request-github",
+				"naumovs.color-highlight",
+				"streetsidesoftware.code-spell-checker",
+				"patr-cloud.patr-theme",
+				"bradlc.vscode-tailwindcss"
+			]
+		}
+	},
+	"forwardPorts": [
+		3000,
+		3001,
+		4000,
+		"logs:3100",
+		"tracing:16686",
+		"grafana:3000"
+	],
+	"postAttachCommand": "sudo chown :docker /var/run/docker.sock || exit 0"
 }
