--- config/tailwind.config.js
@@ -1,12 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-    module.exports = {
-      content: {
-        relative: true,
-        files: ["*.html", "./src/**/*.rs"],
-      },
-      theme: {
-        extend: {},
-      },
-      plugins: [],
-    }
-
